requirements.txt
pandas
matplotlib
seaborn
streamlit
python-docx
pydantic
openpyxl
llama-cloud-services
